refactor: Change Package Structure
최상위에 각 도메인이 존재하고 내부에 UserController면 user패키지의 controller 패키지 내부에 위치시키는 식으로 변경

그냥 User 클래스의 경우 뒤에 붙은게 없으니 user 패키지에 바로 위치 시키는 식
