+++ server/src/main/java/net/chatfoodie/server/chatroom/repository/ChatroomRepository.java
@@ -1,6 +1,6 @@
-package net.chatfoodie.server.domain.chatroom.repository;
+package net.chatfoodie.server.chatroom.repository;
 
-import net.chatfoodie.server.domain.chatroom.entity.Chatroom;
+import net.chatfoodie.server.chatroom.Chatroom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChatroomRepository extends JpaRepository<Chatroom, Long> {
