+++ server/src/main/java/net/chatfoodie/server/message/repository/MessageRepository.java
@@ -1,6 +1,6 @@
-package net.chatfoodie.server.domain.message.repository;
+package net.chatfoodie.server.message.repository;
 
-import net.chatfoodie.server.domain.message.entity.Message;
+import net.chatfoodie.server.message.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MessageRepository extends JpaRepository<Message, Long> {
